Move data filters into dedicated module
